Break out nested classes in Control.ActiveXImpl.
Also do some code cleanup in the main class.

Added a couple sanity tests on VARIANT behavior with the .NET runtime.
